mentalhealthchatbotbackend-main/mentalhealthchatbotbackend-main/pyproject.toml
[project]
name = "mentalhealthbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi>=0.116.1",
    "google-genai>=1.28.0",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "uvicorn[standard]>=0.35.0",
]
